criação de um botão personalizado que muda a cor de fundo
